rename Lift to Symbolic
