+++ DifferentiableDouble/src/test/scala/com/thoughtworks/deeplearning/LiftDoubleSpec.scala
@@ -3,7 +3,7 @@ package com.thoughtworks.deeplearning
 import cats._
 import cats.implicits._
 import com.thoughtworks.deeplearning.DifferentiableDouble._
-import com.thoughtworks.deeplearning.Lift._
+import com.thoughtworks.deeplearning.Symbolic._
 import com.thoughtworks.deeplearning.Layer.Batch
 import com.thoughtworks.deeplearning._
 import org.scalatest._
