+++ DifferentiableDouble/src/test/scala/com/thoughtworks/deeplearning/ToLayerSpec.scala
@@ -1,7 +1,7 @@
 package com.thoughtworks.deeplearning
 
 import cats._
-import com.thoughtworks.deeplearning.Lift._
+import com.thoughtworks.deeplearning.Symbolic._
 import com.thoughtworks.deeplearning.DifferentiableDouble._
 import org.scalatest.{FreeSpec, Matchers}
