+++ DifferentiableINDArray/src/test/scala/com/thoughtworks/deeplearning/DifferentiableINDArraySpec.scala
@@ -1,7 +1,7 @@
 package com.thoughtworks.deeplearning
 
 import cats.Eval
-import com.thoughtworks.deeplearning.Lift._
+import com.thoughtworks.deeplearning.Symbolic._
 import com.thoughtworks.deeplearning.DifferentiableINDArray._
 import com.thoughtworks.deeplearning.DifferentiableINDArray.Optimizers.LearningRate
 import com.thoughtworks.deeplearning.DifferentiableDouble._
