+++ DifferentiableInt/src/main/scala/com/thoughtworks/deeplearning/DifferentiableInt.scala
@@ -3,7 +3,7 @@ package com.thoughtworks.deeplearning
 import cats._
 import cats.implicits._
 import com.thoughtworks.deeplearning.Layer.Batch
-import com.thoughtworks.deeplearning.Lift._
+import com.thoughtworks.deeplearning.Symbolic._
 import com.thoughtworks.deeplearning.DifferentiableAny._
 import com.thoughtworks.deeplearning.Poly.MathMethods
 import com.thoughtworks.deeplearning.Poly.MathMethods./
