+++ DifferentiableSeq/src/main/scala/com/thoughtworks/deeplearning/DifferentiableSeq.scala
@@ -1,9 +1,9 @@
 package com.thoughtworks.deeplearning
 
 import com.thoughtworks.deeplearning.Layer.{Batch, CloseableOnce}
-import com.thoughtworks.deeplearning.Lift._
+import com.thoughtworks.deeplearning.Symbolic._
 import com.thoughtworks.deeplearning.DifferentiableSeq.Layers.{Get, ToSeq}
-import com.thoughtworks.deeplearning.Lift.Layers.Literal
+import com.thoughtworks.deeplearning.Symbolic.Layers.Literal
 import shapeless.Lazy
 
 import language.implicitConversions
