+++ src/test/scala/com/thoughtworks/deeplearning/FortuneTeller.scala
@@ -10,7 +10,7 @@ import com.thoughtworks.deeplearning.DifferentiableBoolean._
 import com.thoughtworks.deeplearning.DifferentiableNothing._
 import com.thoughtworks.deeplearning.DifferentiableDouble._
 import com.thoughtworks.deeplearning.DifferentiableSeq._
-import com.thoughtworks.deeplearning.Lift._
+import com.thoughtworks.deeplearning.Symbolic._
 import com.thoughtworks.deeplearning.DifferentiableCoproduct._
 import org.scalatest._
 import com.thoughtworks.deeplearning.Poly.MathOps
