+++ src/test/scala/com/thoughtworks/deeplearning/MaxSpec.scala
@@ -1,7 +1,7 @@
 package com.thoughtworks
 package deeplearning
 
-import com.thoughtworks.deeplearning.Lift._
+import com.thoughtworks.deeplearning.Symbolic._
 import com.thoughtworks.deeplearning.DifferentiableHList._
 import com.thoughtworks.deeplearning.DifferentiableDouble._
 import com.thoughtworks.deeplearning.DifferentiableAny._
